% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultOrd.R
\name{ctrl.multord}
\alias{ctrl.multord}
\title{Control Function for multord}
\usage{
ctrl.multord(RS = TRUE, thresholds.acat = c("full", "shift",
  "minimal"), XforRS = TRUE, opt.method = c("L-BFGS-B", "nlminb"),
  Q = 10, cores = 5, lambda = 0.01)
}
\arguments{
\item{RS}{Logical value indicating whether response style should be modelled.}

\item{thresholds.acat}{Type of parametrization used for thresholds: \code{thresholds = "full"} implies
separate estimates of threshold values for each response variable; \code{thresholds = "shift"} implies
equal threshold parameter across all response variables modified by shift parameters for
each response variable; \code{thresholds = "minimal"} implies equal threshold parameter across all response variables. This option only applies
for adjacent categories models (\code{model = "acat"} and is not implemented for cumulative models.)}

\item{XforRS}{Logical value indicating whether also covariate effects on the
response style should be considered. Only relevant if \code{RS = TRUE}.}

\item{opt.method}{Specifies optimization algorithm used by \code{\link{optim}}, either
\code{L-BFGS-B} or \code{nlminb}.}

\item{Q}{Number of nodes to be used (per dimension) in Gauss-Hermite-Quadrature. If \code{RS = TRUE},
Gauss-Hermite-Quadrature is two-dimensional.}

\item{cores}{Number of cores to be used in parallelized computation.}

\item{lambda}{Tuning parameter for internal ridge penalty. It is supposed to be set to a small value
to stabilize estimates.}
}
\description{
Control function for multord, a model for multivariate ordinal responses
response styles
}
\examples{
\dontrun{
data(confidence)

## create a small subset of the data to speed up calculations
set.seed(1860)
confidence <- confidence[sample(1:nrow(confidence), 300),]

## scale all metric variables to get comparable parameter estimates
confidence$age <- scale(confidence$age)
confidence$income <- scale(confidence$income)

## two formulas, one without and one with explanatory variables (gender and age)
f.conf0 <- as.formula(paste("cbind(",paste(names(confidence)[1:4],collapse=","),") ~ 1"))
f.conf1 <- as.formula(paste("cbind(",paste(names(confidence)[1:4],collapse=","),") ~ gender + age"))



####
## Adjacent Categories Models
####

## Multivariate adjacent categories model, without response style, without explanatory variables
m.conf0 <- multord(f.conf0, data = confidence, control = ctrl.multord(RS = FALSE))
m.conf0

## Multivariate adjacent categories model, with response style as a random effect, without explanatory variables
m.conf1 <- multord(f.conf0, data = confidence)
m.conf1

## Multivariate adjacent categories model, with response style as a random effect, 
## without explanatory variables for response style BUT for location
m.conf2 <- multord(f.conf1, data = confidence, control = ctrl.multord(XforRS = FALSE))
m.conf2

## Multivariate adjacent categories model, with response style as a random effect, with explanatory variables for location AND response style
m.conf3 <- multord(f.conf1, data = confidence)
m.conf3

plot(m.conf3)



####
## Cumulative Models
####

## Multivariate cumulative model, without response style, without explanatory variables
m.conf0.cumul <- multord(f.conf0, data = confidence, control = ctrl.multord(RS = FALSE), model = "cumulative")
m.conf0.cumul

## Multivariate cumulative model, with response style as a random effect, without explanatory variables
m.conf1.cumul <- multord(f.conf0, data = confidence, model = "cumulative")
m.conf1.cumul

## Multivariate cumulative model, with response style as a random effect, 
## without explanatory variables for response style BUT for location
m.conf2.cumul <- multord(f.conf1, data = confidence, control = ctrl.multord(XforRS = FALSE), model = "cumulative")
m.conf2.cumul

## Multivariate cumulative model, with response style as a random effect, with explanatory variables for location AND response style
m.conf3.cumul <- multord(f.conf1, data = confidence, model = "cumulative")
m.conf3.cumul

plot(m.conf3.cumul)
}
}
\seealso{
\code{\link{multord}} \code{\link{MultOrd-package}} \code{\link{plot.MultOrd}}
}
\author{
Gunther Schauberger\cr \email{gunther.schauberger@tum.de}\cr
\url{https://www.researchgate.net/profile/Gunther_Schauberger2}
}
\keyword{adjacent}
\keyword{categories}
\keyword{cumulative}
\keyword{multivariate}
\keyword{ordinal}
\keyword{response}
\keyword{style}
