% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multord.R
\name{plot.MultOrd}
\alias{plot.MultOrd}
\title{Plot function for MultOrd}
\usage{
\method{plot}{MultOrd}(x, alpha = 0.05, CIfactor = 0.9, \dots)
}
\arguments{
\item{x}{\code{MultOrd} object}

\item{alpha}{Specifies the confidence level 1-alpha of the confidence interval.}

\item{...}{Further plot arguments.}

\item{KI.factor}{Argument that helps to control the appearance (the width) of the stars that represent the confidence intervals of both 
paraemters (location and response style) corresponding to one covariate.}
}
\description{
Plot function for a \code{MultOrd} object. Plots show coefficients of the explanatory variables, both with repect to location and response styles.
The coefficient pairs are displayed as stars, where the rays represent (1-alpha) confidence intervals.
}
\examples{
\dontrun{
data(confidence)

## create a small subset of the data to speed up calculations
set.seed(1860)
confidence <- confidence[sample(1:nrow(confidence), 300),]

## scale all metric variables to get comparable parameter estimates
confidence$age <- scale(confidence$age)
confidence$income <- scale(confidence$income)

## two formulas, one without and one with explanatory variables (gender and age)
f.conf0 <- as.formula(paste("cbind(",paste(names(confidence)[1:4],collapse=","),") ~ 1"))
f.conf1 <- as.formula(paste("cbind(",paste(names(confidence)[1:4],collapse=","),") ~ gender + age"))



####
## Adjacent Categories Models
####

## Multivariate adjacent categories model, without response style, without explanatory variables
m.conf0 <- multord(f.conf0, data = confidence, control = ctrl.multord(RS = FALSE))
m.conf0

## Multivariate adjacent categories model, with response style as a random effect, without explanatory variables
m.conf1 <- multord(f.conf0, data = confidence)
m.conf1

## Multivariate adjacent categories model, with response style as a random effect, 
## without explanatory variables for response style BUT for location
m.conf2 <- multord(f.conf1, data = confidence, control = ctrl.multord(XRS = FALSE))
m.conf2

## Multivariate adjacent categories model, with response style as a random effect, with explanatory variables for location AND response style
m.conf3 <- multord(f.conf1, data = confidence)
m.conf3

plot(m.conf3)



####
## Cumulative Models
####

## Multivariate cumulative model, without response style, without explanatory variables
m.conf0.cumul <- multord(f.conf0, data = confidence, control = ctrl.multord(RS = FALSE), model = "cumulative")
m.conf0.cumul

## Multivariate cumulative model, with response style as a random effect, without explanatory variables
m.conf1.cumul <- multord(f.conf0, data = confidence, model = "cumulative")
m.conf1.cumul

## Multivariate cumulative model, with response style as a random effect, 
## without explanatory variables for response style BUT for location
m.conf2.cumul <- multord(f.conf1, data = confidence, control = ctrl.multord(XRS = FALSE), model = "cumulative")
m.conf2.cumul

## Multivariate cumulative model, with response style as a random effect, with explanatory variables for location AND response style
m.conf3.cumul <- multord(f.conf1, data = confidence, model = "cumulative")
m.conf3.cumul

plot(m.conf3.cumul)
}
}
\seealso{
\code{\link{multord}}, \code{\link{ctrl.multord}}
}
\author{
Gunther Schauberger\cr \email{gunther.schauberger@tum.de}\cr
\url{https://www.researchgate.net/profile/Gunther_Schauberger2}
}
