% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultOrd.R
\name{ctrl.multord}
\alias{ctrl.multord}
\title{Control Function for multord}
\usage{
ctrl.multord(RS = TRUE, thresholds.acat = c("full", "shift",
  "minimal"), XforRS = TRUE, opt.method = c("L-BFGS-B", "nlminb"),
  Q = 10, cores = 5, lambda = 0.01)
}
\arguments{
\item{RS}{Logical value indicating whether response style should be modelled.}

\item{thresholds.acat}{Type of parametrization used for thresholds: \code{thresholds = "full"} implies
separate estimates of threshold values for each response variable; \code{thresholds = "shift"} implies
equal threshold parameter across all response variables modified by shift parameters for
each response variable; \code{thresholds = "minimal"} implies equal threshold parameter across all response variables. This option only applies
for adjacent categories models (\code{model = "acat"} and is not implemented for cumulative models.)}

\item{XforRS}{Logical value indicating whether also covariate effects on the
response style should be considered. Only relevant if \code{RS = TRUE}.}

\item{opt.method}{Specifies optimization algorithm used by \code{\link{optim}}, either
\code{L-BFGS-B} or \code{nlminb}.}

\item{Q}{Number of nodes to be used (per dimension) in Gauss-Hermite-Quadrature. If \code{RS = TRUE},
Gauss-Hermite-Quadrature is two-dimensional.}

\item{cores}{Number of cores to be used in parallelized computation.}

\item{lambda}{Tuning parameter for internal ridge penalty. It is supposed to be set to a small value
to stabilize estimates.}
}
\description{
Control function for multord, a model for multivariate ordinal responses
response styles
}
\examples{
\dontshow{
 }
}
\seealso{
\code{\link{multord}} \code{\link{MultOrd-package}} \code{\link{plot.MultOrd}}
}
\author{
Gunther Schauberger\cr \email{gunther.schauberger@tum.de}\cr
\url{https://www.researchgate.net/profile/Gunther_Schauberger2}
}
\keyword{adjacent}
\keyword{categories}
\keyword{cumulative}
\keyword{multivariate}
\keyword{ordinal}
\keyword{response}
\keyword{style}
